requirements.txt
streamlit
google-genai
python-dotenv
pandas
fpdf
requests
pillow 